Wait for NCI notification before completing the Reset/ModeSet sequences.
In NCI2.0 for commands like CORE_RESET_CMD and NFCEE_MODE_SET_CMD completion
criteria is a post-Response Notification, unlike NCI1x where it was just a
Response message. In some command sequences in NfcCx those Notifications are
not being waited for. It leads to the scenarios when subsequent commands
sent to NFCEE can fail, if sent after Reset/ModeSet response but before the
actual completion notification.