Remove 'NumDriverRestarts' reg-key value from NfcCx.
WDF currently has a bug where it is possible for UMDF drivers to create
sub-keys under the device's registry key. This was not supposed to be
possible and will be fixed sometime soon.

NfcCx is using this bug to create a volatile registry key. That is, a
registry key that is reset after a OS reboot. This is used to store the
number of device restarts that have occured since boot.

When a failure occurs, NfcCx tries to restart the driver up to 5 times
before giving up. But this count is reset when the OS reboots.

However it is no longer possible to implement this feature. So NfcCx
will now instead just give up after the first failure.