UnRegister callback on HCI ClearAllPipe notification #46
During HCI network initialization, secure elements events callbacks are
getting registered in phLibNfc_HciGetSessionIdentityProc.
For some reasons, it may happen that one of the connected secure element
decides to regenerate its pipes. In this case, one of the first HCI
notification will be a clearallpipes, which will requires to reset the
listed pipe for the given secure element as well as the events callback.
One of the use case which may generate this kind of situation might be a
secure element firmware update.